+++ extensions/authclient/clients/VKontakte.php
@@ -80,11 +80,11 @@ class VKontakte extends OAuth2
     /**
      * @inheritdoc
      */
-    protected function apiInternal($accessToken, $url, $method, array $params)
+    protected function apiInternal($accessToken, $url, $method, array $params, array $headers)
     {
         $params['uids'] = $accessToken->getParam('user_id');
         $params['access_token'] = $accessToken->getToken();
-        return $this->sendRequest($method, $url, $params);
+        return $this->sendRequest($method, $url, $params, $headers);
     }
 
     /**