+++ extensions/authclient/CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 authclient extension Change Log
 
 - Bug #3633: OpenId return URL comparison advanced to prevent url encode problem (klimov-paul)
 - Enh #3416: VKontakte OAuth support added (klimov-paul)
+- Enh #4076: Request HTTP headers argument added to `yii\authclient\BaseOAuth::api()` method (klimov-paul)
 - Enh #4139: User attributes requesting at GoogleOAuth switched to Google+ API (klimov-paul)